[proxima-direct-io-pubsub] #333 Add gps-bulk scheme for io-pubsub
