+++ direct/io-pubsub/module/src/test/java/cz/o2/proxima/direct/io/pubsub/PubSubBulkWriterTest.java
@@ -15,35 +15,38 @@
  */
 package cz.o2.proxima.direct.io.pubsub;
 
+import static cz.o2.proxima.direct.io.pubsub.PubSubBulkReader.inflate;
 import static org.junit.Assert.*;
-import static org.junit.Assert.assertTrue;
 
 import com.google.cloud.pubsub.v1.Publisher;
 import com.google.pubsub.v1.PubsubMessage;
 import cz.o2.proxima.core.storage.StreamElement;
 import cz.o2.proxima.core.util.Optionals;
-import cz.o2.proxima.core.util.TestUtils;
-import cz.o2.proxima.direct.core.AttributeWriterBase;
 import cz.o2.proxima.direct.core.Context;
+import cz.o2.proxima.io.pubsub.proto.PubSub.Bulk;
+import cz.o2.proxima.io.pubsub.proto.PubSub.BulkWrapper;
+import cz.o2.proxima.io.pubsub.proto.PubSub.BulkWrapper.Compression;
 import cz.o2.proxima.io.pubsub.util.PubSubUtils;
 import java.io.IOException;
 import java.net.URI;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.UUID;
 import java.util.function.Consumer;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.Test;
+import org.junit.Before;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.junit.runners.Parameterized.Parameter;
+import org.junit.runners.Parameterized.Parameters;
 
-/** Test suite for {@link PubSubWriter}. */
-@Slf4j
-public class PubSubWriterTest extends AbstractPubSubWriterTest {
+@RunWith(Parameterized.class)
+public class PubSubBulkWriterTest extends AbstractPubSubWriterTest {
 
-  private static class TestPubSubWriter extends PubSubWriter implements TestWriter {
+  private static class TestPubSubBulkWriter extends PubSubBulkWriter implements TestWriter {
 
     private Consumer<PubsubMessage> consumer;
 
-    public TestPubSubWriter(PubSubAccessor accessor, Context context) {
+    public TestPubSubBulkWriter(PubSubAccessor accessor, Context context) {
       super(accessor, context);
     }
 
@@ -65,57 +68,66 @@ public class PubSubWriterTest extends AbstractPubSubWriterTest {
     }
   }
 
-  private final PubSubAccessor accessor;
-
-  public PubSubWriterTest() {
-    accessor =
-        new PubSubAccessor(
-            storage, entity, URI.create("gps://my-project/topic"), Collections.emptyMap());
+  @Parameters
+  public static List<Boolean> params() {
+    return Arrays.asList(false, true);
   }
 
-  @Test
-  public void testAsFactorySerializable() throws IOException, ClassNotFoundException {
-    byte[] bytes = TestUtils.serializeObject(writer.asFactory());
-    AttributeWriterBase.Factory<?> factory = TestUtils.deserializeObject(bytes);
-    assertEquals(accessor.getUri(), factory.apply(repo).getUri());
+  @Parameter public boolean deflate;
+
+  private PubSubAccessor accessor;
+
+  @Before
+  public void setUp() {
+    accessor = new PubSubAccessor(storage, entity, getUri(), Collections.emptyMap());
+    super.setUp();
+  }
+
+  private URI getUri() {
+    return deflate
+        ? URI.create("gps-bulk://my-project/topic?deflate=true")
+        : URI.create("gps-bulk://my-project/topic");
   }
 
   @Override
-  public TestWriter createWriter() {
-    return new TestPubSubWriter(accessor, context);
+  TestWriter createWriter() {
+    return new TestPubSubBulkWriter(accessor, context);
   }
 
   @Override
-  void validateWrittenElements(List<PubsubMessage> written, long now) {
-    assertEquals(3, written.size());
-
+  void validateWrittenElements(List<PubsubMessage> written, long now) throws IOException {
+    assertEquals(1, written.size());
+    Bulk bulk = Bulk.parseFrom(maybeInflate(written.get(0).getData().toByteArray()));
+    assertEquals(3, bulk.getKvCount());
+    assertEquals(3, bulk.getUuidCount());
     StreamElement elem =
-        Optionals.get(
-            PubSubUtils.toStreamElement(
-                entity, UUID.randomUUID().toString(), written.get(0).getData().toByteArray()));
+        Optionals.get(PubSubUtils.toStreamElement(entity, bulk.getUuid(0), bulk.getKv(0)));
     assertEquals("key1", elem.getKey());
     assertEquals("attr", elem.getAttribute());
     assertFalse(elem.isDelete());
     assertFalse(elem.isDeleteWildcard());
     assertArrayEquals(new byte[] {1, 2}, elem.getValue());
     assertEquals(now, elem.getStamp());
-    elem =
-        Optionals.get(
-            PubSubUtils.toStreamElement(
-                entity, UUID.randomUUID().toString(), written.get(1).getData().toByteArray()));
+    elem = Optionals.get(PubSubUtils.toStreamElement(entity, bulk.getUuid(1), bulk.getKv(1)));
     assertEquals("key2", elem.getKey());
     assertEquals("attr", elem.getAttribute());
     assertTrue(elem.isDelete());
     assertFalse(elem.isDeleteWildcard());
     assertEquals(now + 1000L, elem.getStamp());
-    elem =
-        Optionals.get(
-            PubSubUtils.toStreamElement(
-                entity, UUID.randomUUID().toString(), written.get(2).getData().toByteArray()));
+    elem = Optionals.get(PubSubUtils.toStreamElement(entity, bulk.getUuid(2), bulk.getKv(2)));
     assertEquals("key3", elem.getKey());
     assertEquals(wildcard.toAttributePrefix() + "*", elem.getAttribute());
     assertTrue(elem.isDelete());
     assertTrue(elem.isDeleteWildcard());
     assertEquals(now, elem.getStamp());
+  }
+
+  private byte[] maybeInflate(byte[] bytes) throws IOException {
+    BulkWrapper wrapper = BulkWrapper.parseFrom(bytes);
+    assertEquals(deflate ? Compression.DEFLATE : Compression.NONE, wrapper.getCompression());
+    if (deflate) {
+      return inflate(wrapper.getValue().toByteArray());
+    }
+    return wrapper.getValue().toByteArray();
   }
 }
